% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverFunction.R
\name{colPlot}
\alias{colPlot}
\title{colPlot function}
\usage{
colPlot(data, date, color, varName, title)
}
\arguments{
\item{data}{Input dataset.}

\item{date}{End date.}

\item{color}{Color of the columns in the plot.}

\item{varName}{Variable name. The variable name must be within the dataset --- \code{data}.}

\item{title}{Title of the col plot.}
}
\description{
This function is used to create column plot.
}
