% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\docType{data}
\name{covid}
\alias{covid}
\title{Global covid-19 cases and deaths}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68150 rows and 8 columns.
}
\source{
World Health Organisation (WHO)
}
\usage{
covid
}
\description{
This dataset contains 225 countries
}
\keyword{datasets}
