% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19ViewerShiny-package.R
\docType{package}
\name{covid19ViewerShiny-package}
\alias{covid19ViewerShiny}
\alias{covid19ViewerShiny-package}
\title{covid19ViewerShiny: A Shiny App Visualising Global Covid19 Cases And Deaths}
\description{
This package contains WHO covid dataset and a Shiny app that can monitor global covid-19 cases and compare the cases by countries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-ycui0008}
  \item Report bugs at \url{https://github.com/etc5523-2020/r-package-assessment-ycui0008/issues}
}

}
\author{
\strong{Maintainer}: Yuheng Cui \email{ycui0008@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
