% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launch_app}
\alias{launch_app}
\title{Run \verb{covid-19 Viewer} Shiny app}
\usage{
launch_app()
}
\description{
Runs \verb{covid-19 Viewer} Shiny app contained in the \code{covid19ViewerShiny} package.
}
