% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiFunction.R
\name{multiSelection}
\alias{multiSelection}
\title{Create a multiSelection drop list}
\usage{
multiSelection(id, inputDisplay = " ", selection = covid$country)
}
\arguments{
\item{id}{id name of the input, it must be a character}

\item{inputDisplay}{The name of the selection drop list}

\item{selection}{A list of character.}
}
\description{
This function can help user to create a drop list in the Shiny app.
}
\examples{
multiSelection("country_select1", "Country:", covid$country)

}
