% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{runShinyPackageApp}
\alias{runShinyPackageApp}
\title{Run \verb{covid-19 Viewer} Shiny app}
\usage{
runShinyPackageApp()
}
\description{
Runs \verb{covid-19 Viewer} Shiny app contained in the \code{covid19ViewerShiny} package.
}
