% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverFunction.R
\name{colPlot}
\alias{colPlot}
\title{colPlot function}
\usage{
colPlot(data = covid, date = input$end_date2, color = "red", varName, title)
}
\arguments{
\item{data}{Input dataset.}

\item{date}{End date.}

\item{color}{Color of the columns in the plot.}

\item{varName}{Variable name. Recommend user enter either \code{cumulative_cases} or \code{cumulative_deaths} here.}

\item{title}{Optional. Title of the col plot.}
}
\description{
This function is used to create column plot.
}
